--- Assets/Internal/Script/Player/PlayerAttack2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 941265b1a424f1d4a92f408767df1d24
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
